FDS-115 Updated very old Form POC with actual components and modules from Table
This also includes adding the concept of component-specific modules. Form now has a data module for rendering rows.

We will need to bring back the old ActionEdit button… but possibly make that specific to Form. This pattern of merging the module types to check them when throwing prop errors is something we should evaluate in both Form and Table.